+++ fastpyxl/worksheet/tests/test_read_only.py
@@ -17,6 +17,7 @@ def DummyWorkbook():
         epoch = None
         _cell_styles = [StyleArray([0, 0, 0, 0, 0, 0, 0, 0, 0])]
         data_only = False
+        keep_formula_cache = False
 
         def __init__(self):
             self.sheetnames = []
@@ -74,7 +75,7 @@ class TestReadOnlyWorksheet:
 
     def test_empty_cell(self, ReadOnlyWorksheet):
         row = [
-            (1, 4, None, 'n', 0),
+            (1, 4, None, 'n', 0, None),
         ]
         ws = ReadOnlyWorksheet
         cells = ws._get_row(row, max_col=4, values_only=True)
@@ -83,8 +84,8 @@ class TestReadOnlyWorksheet:
 
     def test_pad_row_left(self, ReadOnlyWorksheet):
         row = [
-            (1, 4, 4, 'n', 0),
-            (1, 8, 8, 'n', 0),
+            (1, 4, 4, 'n', 0, None),
+            (1, 8, 8, 'n', 0, None),
         ]
         ws = ReadOnlyWorksheet
         cells = ws._get_row(row, max_col=4, values_only=True)
@@ -93,8 +94,8 @@ class TestReadOnlyWorksheet:
 
     def test_pad_row(self, ReadOnlyWorksheet):
         row = [
-            (1, 4, 4, 'n', 0),
-            (1, 8, 8, 'n', 0),
+            (1, 4, 4, 'n', 0, None),
+            (1, 8, 8, 'n', 0, None),
         ]
         ws = ReadOnlyWorksheet
         cells = ws._get_row(row, min_col=4, max_col=8, values_only=True)
@@ -103,8 +104,8 @@ class TestReadOnlyWorksheet:
 
     def test_pad_row_right(self, ReadOnlyWorksheet):
         row = [
-            (1, 4, 4, 'n', 0),
-            (1, 8, 8, 'n', 0),
+            (1, 4, 4, 'n', 0, None),
+            (1, 8, 8, 'n', 0, None),
         ]
         ws = ReadOnlyWorksheet
         cells = ws._get_row(row, min_col=6, max_col=10, values_only=True)
@@ -113,8 +114,8 @@ class TestReadOnlyWorksheet:
 
     def test_pad_row_cells(self, ReadOnlyWorksheet):
         row = [
-            (2, 4, 4, 'n', 0),
-            (2, 8, 8, 'n', 0),
+            (2, 4, 4, 'n', 0, None),
+            (2, 8, 8, 'n', 0, None),
         ]
         ws = ReadOnlyWorksheet
         cells = ws._get_row(row, min_col=6, max_col=10)
@@ -226,6 +227,7 @@ def test_implementation_compatbility(ReadOnlyWorksheet, DummyWorkbook):
     std_only = set(['HeaderFooter',
                     '_WorkbookChild__title',
                     '_cells',
+                    '_formula_caches',
                     '_charts',
                     '_comments',
                     '_current_row',